pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = "educational project"
authors = ["Alexey.S <Morphxoxo@gmail.com>"]
readme = "README.md"
packages = [
  { include = "gendiff" },
]

[tool.poetry.dependencies]
python = "^3.8"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
pytest = "^7.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
gendiff = "gendiff.scripts.gendiff:main"
